fix: renaming wheelpicker
