+++ packages/trader/src/AppV2/Components/TradeParameters/Duration/duration-wheel-picker.tsx
